feat(buildkit): add the buildkit core module and move Stage into it
The generic Stage[I, O, S] mentioned neither Elm nor Morphir, yet the Elm
compiler API owned it. It now lives in the new cross-platform
morphir/buildkit/core module, which depends only on Kyo and carries no
Morphir domain type. The contract is unchanged.

Part of intent 0008.
