+++ morphir/buildkit/core/src/morphir/buildkit/core/Stage.scala
@@ -1,4 +1,4 @@
-package morphir.langkit.elm.compiler
+package morphir.buildkit.core
 
 import kyo.*
