+++ morphir/buildkit/core/test/src/morphir/buildkit/core/StageSpec.scala
@@ -1,4 +1,4 @@
-package morphir.langkit.elm.compiler
+package morphir.buildkit.core
 
 import kyo.*
 import kyo.test.*
